+++ txbits/app/controllers/IAPI/CryptoAddress.scala
@@ -2,7 +2,7 @@
 // This file is licensed under the Affero General Public License version 3 or later,
 // see the accompanying file COPYING or http://www.gnu.org/licenses/agpl.html.
 
-package controllers.API
+package controllers.IAPI
 
 import com.google.bitcoin.core.Address
 import com.google.bitcoin.core.NetworkParameters
