+++ txbits/public/js/api1.js
@@ -1,5 +1,6 @@
 var API;
 (function(){
+    var prefix = '/iapi/1/';
     // when you call API wrap on a function it returns a function which
     // when called acts just like the function you wrapped but also adds
     // an error handler
@@ -24,19 +25,19 @@ var API;
     };
     API = {
         balance: APIWrap(function() {
-            return $.get('/api/1/balance', 'json');
+            return $.get(prefix+'balance', 'json');
         }),
 
         pending_withdrawals: APIWrap(function(){
-            return $.get('/api/1/pending_withdrawals_all', 'json')
+            return $.get(prefix+'pending_withdrawals_all', 'json')
         }),
 
         pending_deposits: APIWrap(function(){
-            return $.get('/api/1/pending_deposits_all', 'json')
+            return $.get(prefix+'pending_deposits_all', 'json')
         }),
 
         bid: APIWrap(function(base, counter, amount, price) {
-            return $.ajax('/api/1/bid', {
+            return $.ajax(prefix+'bid', {
                 type: 'POST',
                 data: JSON.stringify({base: base, counter: counter, amount: amount, price: price}),
                 dataType: 'json',
@@ -45,7 +46,7 @@ var API;
         }),
 
         ask: APIWrap(function(base, counter, amount, price) {
-            return $.ajax('/api/1/ask', {
+            return $.ajax(prefix+'ask', {
                 type: 'POST',
                 data: JSON.stringify({base: base, counter: counter, amount: amount, price: price}),
                 dataType: 'json',
@@ -54,7 +55,7 @@ var API;
         }),
 
         cancel: APIWrap(function(id) {
-            return $.ajax('/api/1/cancel', {
+            return $.ajax(prefix+'cancel', {
                 type: 'POST',
                 data: JSON.stringify({order: Number(id)}),
                 dataType: 'json',
@@ -63,7 +64,7 @@ var API;
         }),
 
         withdraw: APIWrap(function(currency, amount, address, tfa_code) {
-            return $.ajax('/api/1/withdraw', {
+            return $.ajax(prefix+'withdraw', {
                 type: 'POST',
                 data: JSON.stringify({currency: currency, amount: amount, address: address, tfa_code: tfa_code}),
                 dataType: 'json',
@@ -72,71 +73,71 @@ var API;
         }),
 
         pairs: APIWrap(function() {
-            return $.get('/api/1/pairs', 'json');
+            return $.get(prefix+'pairs', 'json');
         }),
 
         currencies: APIWrap(function() {
-            return $.get('/api/1/currencies', 'json');
+            return $.get(prefix+'currencies', 'json');
         }),
 
         deposit_crypto: APIWrap(function(currency) {
-            return $.get('/api/1/deposit_crypto/'+currency, 'json');
+            return $.get(prefix+'deposit_crypto/'+currency, 'json');
         }),
 
         deposit_crypto_all: APIWrap(function() {
-            return $.get('/api/1/deposit_crypto_all', 'json');
+            return $.get(prefix+'deposit_crypto_all', 'json');
         }),
 
         ticker: APIWrap(function() {
-            return $.get('/api/1/ticker', 'json');
+            return $.get(prefix+'ticker', 'json');
         }),
 
         trade_history: APIWrap(function() {
-            return $.get('/api/1/trade_history', 'json');
+            return $.get(prefix+'trade_history', 'json');
         }),
 
         login_history: APIWrap(function() {
-            return $.get('/api/1/login_history', 'json');
+            return $.get(prefix+'login_history', 'json');
         }),
 
         deposit_withdraw_history: APIWrap(function() {
-            return $.get('/api/1/deposit_withdraw_history', 'json');
+            return $.get(prefix+'deposit_withdraw_history', 'json');
         }),
 
         pending_trades: APIWrap(function() {
-            return $.get('/api/1/pending_trades', 'json');
+            return $.get(prefix+'pending_trades', 'json');
         }),
 
         trade_fees: APIWrap(function() {
-            return $.get('/api/1/trade_fees', 'json');
+            return $.get(prefix+'trade_fees', 'json');
         }),
 
         dw_fees: APIWrap(function() {
-            return $.get('/api/1/dw_fees', 'json');
+            return $.get(prefix+'dw_fees', 'json');
         }),
 
         dw_limits: APIWrap(function() {
-            return $.get('/api/1/dw_limits', 'json');
+            return $.get(prefix+'dw_limits', 'json');
         }),
 
         required_confirms: APIWrap(function() {
-            return $.get('/api/1/required_confirms', 'json');
+            return $.get(prefix+'required_confirms', 'json');
         }),
 
         open_trades: APIWrap(function(first, second) {
-            return $.get('/api/1/open_trades/'+first+'/'+second, 'json');
+            return $.get(prefix+'open_trades/'+first+'/'+second, 'json');
         }),
 
         recent_trades: APIWrap(function(first, second) {
-            return $.get('/api/1/recent_trades/'+first+'/'+second, 'json');
+            return $.get(prefix+'recent_trades/'+first+'/'+second, 'json');
         }),
 
         user: APIWrap(function() {
-            return $.get('/api/1/user', 'json');
+            return $.get(prefix+'user', 'json');
         }),
 
         turnoff_tfa: APIWrap(function(code, password) {
-            return $.ajax('/api/1/turnoff_tfa', {
+            return $.ajax(prefix+'turnoff_tfa', {
                 type: 'POST',
                 data: JSON.stringify({tfa_code: code, password: password}),
                 dataType: 'json',
@@ -145,7 +146,7 @@ var API;
         }),
 
         turnoff_emails: APIWrap(function() {
-            return $.ajax('/api/1/turnoff_emails', {
+            return $.ajax(prefix+'turnoff_emails', {
                 type: 'POST',
                 data: "{}",
                 dataType: 'json',
@@ -154,7 +155,7 @@ var API;
         }),
 
         turnon_emails: APIWrap(function() {
-            return $.ajax('/api/1/turnon_emails', {
+            return $.ajax(prefix+'turnon_emails', {
                 type: 'POST',
                 data: "{}",
                 dataType: 'json',
@@ -163,11 +164,11 @@ var API;
         }),
 
         gen_totp_secret: APIWrap(function() {
-            return $.post('/api/1/gen_totp_secret', 'json');
+            return $.post(prefix+'gen_totp_secret', 'json');
         }),
 
         turnon_tfa: APIWrap(function(code, password) {
-            return $.ajax('/api/1/turnon_tfa', {
+            return $.ajax(prefix+'turnon_tfa', {
                 type: 'POST',
                 data: JSON.stringify({tfa_code: code, password: password}),
                 dataType: 'json',
@@ -176,7 +177,7 @@ var API;
         }),
 
         add_pgp: APIWrap(function(password, code, pgp) {
-            return $.ajax('/api/1/add_pgp', {
+            return $.ajax(prefix+'add_pgp', {
                 type: 'POST',
                 data: JSON.stringify({tfa_code: code, password: password, pgp: pgp}),
                 dataType: 'json',
@@ -185,7 +186,7 @@ var API;
         }),
 
         remove_pgp: APIWrap(function(password, code) {
-            return $.ajax('/api/1/remove_pgp', {
+            return $.ajax(prefix+'remove_pgp', {
                 type: 'POST',
                 data: JSON.stringify({tfa_code: code, password: password}),
                 dataType: 'json',
@@ -194,7 +195,7 @@ var API;
         }),
 
         add_api_key: APIWrap(function() {
-            return $.ajax('/api/1/add_api_key', {
+            return $.ajax(prefix+'add_api_key', {
                 type: 'POST',
                 data: "{}",
                 dataType: 'json',
@@ -203,7 +204,7 @@ var API;
         }),
 
         update_api_key: APIWrap(function(code, api_key, trading, trade_history, list_balance) {
-            return $.ajax('/api/1/update_api_key', {
+            return $.ajax(prefix+'update_api_key', {
                 type: 'POST',
                 data: JSON.stringify({tfa_code: code, api_key: api_key, trading: trading, trade_history: trade_history, list_balance: list_balance}),
                 dataType: 'json',
@@ -212,7 +213,7 @@ var API;
         }),
 
         disable_api_key: APIWrap(function(code, api_key) {
-            return $.ajax('/api/1/disable_api_key', {
+            return $.ajax(prefix+'disable_api_key', {
                 type: 'POST',
                 data: JSON.stringify({tfa_code: code, api_key: api_key}),
                 dataType: 'json',
@@ -221,11 +222,11 @@ var API;
         }),
 
         get_api_keys: APIWrap(function() {
-            return $.get('/api/1/get_api_keys', 'json');
+            return $.get(prefix+'get_api_keys', 'json');
         }),
 
         chart: APIWrap(function(base, counter) {
-            return $.get('/api/1/chart/'+base+'/'+counter, 'json');
+            return $.get(prefix+'chart/'+base+'/'+counter, 'json');
         })
     };
 })();
